initial split __init__
